feat(ui): move heavy-dependency components to dedicated subpath entry points
The tsdown config now uses the same splat entry pattern as @sanity/icons:
every file in packages/ui/src/exports/ becomes an entry point and a
package.json exports subpath, so adding a file is all it takes to publish a
new subpath.

Components that depend on @floating-ui/react-dom, motion or react-refractor
move off the root entry so that importing '@sanity/ui' never references those
dependencies (regardless of bundler treeshaking):

- @sanity/ui/toast: Toast, ToastProvider, useToast
- @sanity/ui/popover: Popover
- @sanity/ui/tooltip: Tooltip, TooltipDelayGroup*
- @sanity/ui/menu: Menu, MenuButton, MenuDivider, MenuGroup, MenuItem
- @sanity/ui/autocomplete: Autocomplete
- @sanity/ui/breadcrumbs: Breadcrumbs
- @sanity/ui/code: Code

The root entry keeps @deprecated never-typed tombstones for every moved
symbol so editors and agents can tell where a symbol went, and ErrorBoundary
no longer renders Code (which would keep react-refractor in the root graph).
