+++ packages/ui/tsconfig.dist.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.settings.json",
-  "include": ["./exports", "./src", "./typings"],
+  "include": ["./src", "./typings"],
   "exclude": [
     "./src/**/__fixtures__",
     "./src/**/__mocks__",
