+++ packages/ui/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "../../tsconfig.settings.json",
   "include": [
-    "./exports",
     "./scripts",
     "./src",
     "./test",
